requirements.txt
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
kafka-python==2.0.2
pyspark==4.0.0
streamlit==1.32.0
plotly==5.19.0
faker==24.0.0